Initial restructuring into a module
